# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: AI CCTV 공개 API — 모바일 교체 계약
  version: 0.3.0
  description: |-
    모바일 등 외부 클라이언트를 교체할 때 유지해야 할 공개 HTTP API이다.
    앱은 중앙 HTTPS origin에 접속하며 Data·Edge·MediaMTX의 내부 포트에 직접 연결하지 않는다.
    paths와 기존 요청·응답 모델은 External의 FastAPI 선언에서 생성한다. 한국어 사용 설명,
    Cookie 인증, Range 헤더 및 Any/dict로 선언된 세 응답은 실제 반환 코드를 근거로 보완한다.
    이 문서는 실행 서버의 /api/v1/openapi.json에 설명을 보강한 배포용 명세다.

    ### 로그인과 권한
    1. POST /api/v1/auth/login으로 access_token·refresh_token과 user를 받는다.
    2. API에는 Authorization: Bearer <access_token> 또는 Access Cookie를 보낸다.
       두 가지가 함께 있으면 Bearer가 우선한다. admin은 전체, viewer는 허용 카메라만 조회한다.
    3. 401이면 현재 refresh_token으로 한 번 갱신하고 원래 요청을 재시도한다.
       갱신은 새 access·refresh 쌍을 반환하며 이전 refresh는 재사용할 수 없다.
       앱은 동시 갱신을 하나로 합치고 최신 두 토큰을 함께 저장한다. 갱신 실패는 재로그인한다.
    4. 로그아웃에는 현재 access와 refresh를 함께 보낸다. 서버는 제출된 두 토큰을 폐기하고
       refresh 계열의 기기 등록·대기 알림과 쿠키를 삭제한다. 다른 access까지 전부 즉시 철회하는
       전체 기기 로그아웃 API는 아니다. 이미 회전된 refresh 대신 반드시 최신값을 보낸다.
    기본 쿠키명은 ai_cctv_access, ai_cctv_refresh이며 배포 설정으로 바뀔 수 있다.
    로그인·갱신은 HttpOnly 쿠키도 설정한다. 기본 Secure=true, SameSite=lax,
    access Path=/, refresh Path=/api/v1/auth, 기본 수명은 각각 900초·604800초이다.
    실제 access 수명은 expires_in을 따른다. 토큰을 URL·로그에 기록하지 않는다.

    ### 영상 재생: Nginx 경로
    GET /cameras/{camera_id}/live의 url 또는 GET /recordings/{id}/playback의 playback_url을
    그대로 사용한다. PUBLIC_BASE_URL 설정 시 HTTPS 절대 URL, 없으면 API origin 기준 상대 URL이다.
    라이브 기본 주소는 /hls/{camera_id}/index.m3u8이고 재생목록과 영상 조각을 받는 HLS이다.
...

...

...

...
paths:
...
  /api/v1/recordings:
    get:
      summary: 녹화 구간 검색
      operationId: list_recordings_api_v1_recordings_get
      security:
      - HTTPBearer: []
      - AccessCookie: []
      parameters:
      - name: camera_id
        in: query
        required: true
        schema:
          type: string
          title: Camera Id
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: From
        description: '시간대를 포함한 RFC 3339. 예: 2026-09-07T00:00:00Z. 함께 지정하면 from < to.'
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: To
        description: '시간대를 포함한 RFC 3339. 예: 2026-09-07T00:00:00Z. 함께 지정하면 from < to.'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
        description: 건너뛸 항목 수. items가 빈 목록이면 조회를 마친다.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ApiError'
      description: camera_id·from·to가 필요하다. 권한이 있는 카메라만 조회한다.
      tags:
      - recordings
  /api/v1/recordings/{segment_id}:
    get:
      summary: 녹화 구간 조회
      operationId: get_recording_api_v1_recordings__segment_id__get
      security:
      - HTTPBearer: []
      - AccessCookie: []
      parameters:
      - name: segment_id
        in: path
        required: true
        schema:
          type: string
          title: Segment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ApiError'
      description: 녹화 정보와 그 카메라의 열람 권한을 확인한다.
      tags:
      - recordings
  /api/v1/recordings/{segment_id}/playback:
    get:
      summary: 녹화 재생 주소
      operationId: get_recording_playback_api_v1_recordings__segment_id__playback_get
      security:
      - HTTPBearer: []
      - AccessCookie: []
      parameters:
      - name: segment_id
        in: path
        required: true
        schema:
          type: string
          title: Segment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPlaybackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ApiError'
      description: format=mpegts면 공개 /content 주소, 그 외에는 MediaMTX fmp4 Playback 주소를 반환한다. 주소를 직접 조합하거나
        모두 HLS라고 가정하지 않는다.
      tags:
      - recordings
  /api/v1/recordings/{segment_id}/content:
    get:
      summary: 녹화 영상 바이트
      operationId: get_recording_content_api_v1_recordings__segment_id__content_get
      security:
      - HTTPBearer: []
      - AccessCookie: []
      parameters:
      - name: segment_id
        in: path
        required: true
        schema:
          type: string
          title: Segment Id
      - name: Range
        in: header
        required: false
        description: '단일 Byte Range. 예: bytes=0-1023, bytes=1024-, bytes=-1024'
        schema:
          type: string
      - name: If-Range
        in: header
        required: false
        description: 이전 응답의 ETag 또는 HTTP 날짜. 일치하지 않으면 전체 200.
        schema:
          type: string
      responses:
        '200':
          description: Complete recording content
          content:
            video/mp2t:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
          headers:
            Accept-Ranges:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Range:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            ETag:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Last-Modified:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Length:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
        '206':
          description: Recording byte range
          content:
            video/mp2t:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
          headers:
            Accept-Ranges:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Range:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            ETag:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Last-Modified:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Length:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
        '416':
          description: Unsatisfiable byte range
          headers:
            Accept-Ranges:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Range:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            ETag:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Last-Modified:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
            Content-Length:
              schema:
                type: string
              description: 상태와 요청에 따라 제공되는 원본 영상 응답 헤더
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ApiError'
      description: Range가 없으면 전체 200, 유효한 단일 Byte Range는 206, 범위를 만족할 수 없으면 416이다. If-Range의 ETag·시각이
        현재 파일과 다르면 전체 200을 반환한다.
      tags:
      - recordings
  /api/v1/recovery-jobs:
    get:
      summary: 영상 복구 작업 목록
      operationId: list_recovery_jobs_api_v1_recovery_jobs_get
      security:
      - HTTPBearer: []
      - AccessCookie: []
      parameters:
      - name: camera_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
        description: 건너뛸 항목 수. items가 빈 목록이면 조회를 마친다.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryJobPageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ApiError'
      description: 관리자 전용. Edge 장애 구간의 복구 진행 상태와 오류를 조회한다.
      tags:
      - recovery-jobs
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RecordingPageResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RecordingResponse'
          type: array
          title: Items
        limit:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Limit
        offset:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Offset
      type: object
      required:
      - items
      title: RecordingPageResponse
    RecordingPlaybackResponse:
      properties:
        recording_id:
          type: string
          title: Recording Id
        playback_url:
          type: string
          title: Playback Url
      type: object
      required:
      - recording_id
      - playback_url
      title: RecordingPlaybackResponse
    RecordingResponse:
      properties:
        id:
          anyOf:
          - type: integer
          - type: string
          title: Id
        camera_id:
          type: string
          pattern: ^[a-z0-9][a-z0-9_-]{0,63}$
          title: Camera Id
        start_time:
          type: string
          format: date-time
          title: Start Time
        end_time:
          type: string
          format: date-time
          title: End Time
        format:
          anyOf:
          - type: string
            enum:
            - fmp4
            - mp4
            - mpegts
          - type: 'null'
          title: Format
        codec:
          anyOf:
          - type: string
          - type: 'null'
          title: Codec
        duration_ms:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Duration Ms
        file_size:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: File Size
        source:
          anyOf:
          - type: string
            enum:
            - central
            - edge_recovery
            - import
          - type: 'null'
          title: Source
        status:
          anyOf:
          - type: string
            enum:
            - writing
            - ready
            - missing
            - corrupt
            - deleting
            - deleted
          - type: 'null'
          title: Status
        checksum:
          anyOf:
          - type: string
          - type: 'null'
          title: Checksum
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - camera_id
      - start_time
      - end_time
      title: RecordingResponse
    RecoveryJobPageResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RecoveryJobResponse'
          type: array
          title: Items
        limit:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Limit
        offset:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Offset
      type: object
      required:
      - items
      title: RecoveryJobPageResponse
    RecoveryJobResponse:
      properties:
        id:
          type: integer
          title: Id
        camera_id:
          type: string
          pattern: ^[a-z0-9][a-z0-9_-]{0,63}$
          title: Camera Id
        outage_started_at:
          type: string
          format: date-time
          title: Outage Started At
        outage_ended_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Outage Ended At
        status:
          type: string
          enum:
          - detected
          - waiting_for_recovery
          - downloading
          - indexing
          - completed
          - failed
          title: Status
        attempt_count:
          type: integer
          minimum: 0.0
          title: Attempt Count
        max_attempts:
          type: integer
          exclusiveMinimum: 0.0
          title: Max Attempts
        revision:
          type: integer
          minimum: 0.0
          title: Revision
          default: 0
        next_retry_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Next Retry At
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        recovery_summary:
          anyOf:
          - type: object
          - type: 'null'
          title: Recovery Summary
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - camera_id
      - outage_started_at
      - outage_ended_at
      - status
      - attempt_count
      - max_attempts
      title: RecoveryJobResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    DetailError:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    ControlError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
          - code
          - message
          - details
      required:
      - error
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access_token. refresh_token을 Bearer로 사용하지 않는다.
    AccessCookie:
      type: apiKey
      in: cookie
      name: ai_cctv_access
      description: 기본 access 쿠키명. 배포의 ACCESS_COOKIE_NAME으로 변경할 수 있다.
  responses:
    ApiError:
      description: 요청·권한·상태·상위 서비스 오류. 일반 detail 또는 제어 error 구조이며 422는 detail 배열이다. 프록시나 처리하지 못한 서버 오류는
        JSON이 아닐 수 있다.
      content:
        application/json:
          schema:
            anyOf:
            - $ref: '#/components/schemas/DetailError'
            - $ref: '#/components/schemas/ControlError'
            - $ref: '#/components/schemas/HTTPValidationError'
        text/plain:
          schema:
            type: string
        text/html:
          schema:
            type: string
    Unauthorized:
      description: 인증 정보가 없거나 만료·철회·변경되었다. 현재 refresh로 한 번만 갱신하고 실패하면 다시 로그인한다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DetailError'
servers:
...